+++ resources/install.rb
@@ -4,7 +4,7 @@ property :git_ref, String,
   default: 'master',
   description: 'Git reference to download, set to a tag to get a specific version'
 
-unified_mode true
+unified_mode true if respond_to? :unified_mode
 
 action :install do
   src_path = "#{Chef::Config['file_cache_path']}/ruby-build"