+++ test/cookbooks/test/recipes/default.rb
@@ -3,4 +3,7 @@ homebrew_update
 
 ruby_build_install
 
-ruby_build_definition '2.6.0'
+ruby_build_definition '2.6.6' do
+  version_prefix true
+  patch 'test.patch'
+end